feat(phase10): complete Sessions 2-3 Protocol Integration
Session 2.4: NAT Traversal Wiring
- Wire NatDetector to Node for NAT type detection
- Add traverse_nat() method with hole punching support
- Implement relay fallback for symmetric NAT
- Add NAT traversal tests

Session 3.1: Crypto to Frames Integration
- Wire SessionCrypto encryption/decryption to frame handling
- Add encrypt_frame() and decrypt_frame() to PeerConnection
- Implement key ratcheting on frame sequence
- Add frame crypto round-trip tests

Session 3.2: File Transfer Wiring
- Add file_transfer.rs with FileTransferManager
- Wire FileChunker to chunk outgoing files
- Add chunk routing through encrypted sessions
- Implement chunk reassembly with BLAKE3 verification
- Add transfer progress tracking

Session 3.3: Obfuscation Wiring
- Wire padding modes to frame sending (PowerOfTwo, SizeClasses)
- Implement timing delays (Fixed, Uniform, Normal, Exponential)
- Add protocol mimicry wrapper (TLS, WebSocket, DoH)
- Implement cover traffic generator with distributions
- Add obfuscation pipeline tests

Session 3.4: Integration Tests
- Add test_transport_initialization
- Add test_encrypted_frame_exchange
- Add test_obfuscation_pipeline
- Add test_file_chunk_transfer
- Add test_cover_traffic_generation
- Add test_discovery_node_integration

Test Results: 1,025+ tests (1,011 passing, 14 ignored for Phase 7)
Quality: Zero clippy warnings, all code formatted

+++ benches/transfer.rs
@@ -169,7 +169,6 @@ fn bench_transfer_throughput(c: &mut Criterion) {
         group.bench_with_input(BenchmarkId::from_parameter(size), &size, |b, &size| {
             b.iter(|| {
                 rt.block_on(async {
-                    use std::fs;
                     use tempfile::NamedTempFile;
                     use wraith_core::node::Node;
 
@@ -239,7 +238,6 @@ fn bench_transfer_latency(c: &mut Criterion) {
     c.bench_function("small_file_transfer_latency", |b| {
         b.iter(|| {
             rt.block_on(async {
-                use std::fs;
                 use tempfile::NamedTempFile;
                 use wraith_core::node::Node;
 
@@ -292,7 +290,7 @@ fn bench_bbr_utilization(c: &mut Criterion) {
                 // Update bandwidth estimation
                 let delivered = 1400 * (i + 1); // MTU-sized packets
                 let interval = std::time::Duration::from_micros(1000);
-                bbr.update_bandwidth(delivered as u64, interval);
+                bbr.update_bandwidth(delivered, interval);
             }
 
             black_box(bbr.cwnd())
@@ -359,7 +357,6 @@ fn bench_multi_peer_speedup(c: &mut Criterion) {
             |b, &num_peers| {
                 b.iter(|| {
                     rt.block_on(async {
-                        use std::fs;
                         use tempfile::NamedTempFile;
                         use wraith_core::node::Node;
